Dashboard V4

+++ app/src/main/java/com/example/vitalage/model/SignoVital.kt
@@ -8,5 +8,5 @@ data class SignoVital(
     val frecuenciaCardiaca: Int = 0,
     val temperatura: Float = 0.0f,
     val oxigenacion: Int = 0,
-    val imc:Float=0.0f,
+    val imc:Double=0.0,
 )
